ai/requirements.txt
pandas==2.3.0
numpy==2.3.0
scikit-learn==1.8.0
sentence-transformers==5.2.0
umap-learn==0.5.9.post2
hdbscan==0.8.41
playwright==1.55.0
matplotlib==3.10.3
seaborn==0.13.2
langdetect==1.0.9
torch==2.7.1
transformers==4.57.3
python-dotenv==1.1.1